Renamed Product IVC to ProductName to avoid conflict with Product Entity, and for clarity
